refactor the mapper
