+++ src/features/applications/mappers/index.ts
@@ -1,10 +1,18 @@
-import { Application, ApplicationGlobalStateType, ApplicationGlobalStateValue, ApplicationSummary } from '../models'
+import {
+  Application,
+  ApplicationGlobalStateType,
+  ApplicationGlobalStateValue,
+  ApplicationSummary,
+  ArgumentHint,
+  MethodDefinition,
+} from '../models'
 import { encodeAddress, getApplicationAddress, modelsv2 } from 'algosdk'
 import isUtf8 from 'isutf8'
 import { Buffer } from 'buffer'
 import { ApplicationMetadataResult, ApplicationResult } from '../data/types'
 import { asJson } from '@/utils/as-json'
 import { Arc32AppSpec } from '@/features/app-interfaces/data/types'
+import algosdk from 'algosdk'
 
 export const asApplicationSummary = (application: ApplicationResult): ApplicationSummary => {
   return {
@@ -35,7 +43,7 @@ export const asApplication = (application: ApplicationResult, metadata: Applicat
     globalState: asGlobalStateValue(application.params['global-state']),
     isDeleted: application.deleted ?? false,
     json: asJson(application),
-    appSpec,
+    methods: appSpec ? asMethodDefinitions(appSpec) : [],
   }
 }
 
@@ -91,38 +99,44 @@ const getValue = (bytes: string) => {
     }
   }
 }
-//
-// export const asAbiMethodDefinition = (abiMethod: algosdk.ABIMethod): AbiMethodDefinition => {
-//   const arguments = abiMethod.args.map((arg) => ({
-//     name: arg.name,
-//     type: arg.type,
-//   }))
-// }
-//
-// const asAbiMethodArgumentDefinition = (arg: algosdk.ABIMethod['args'][0]): AbiMethodArgumentDefinition => {
-//   if (algosdk.abiTypeIsTransaction(arg.type)) {
-//     return {
-//       name: arg.name,
-//       description: arg.description,
-//       type: AbiType.Transaction,
-//     }
-//   }
-//
-//   if (algosdk.abiTypeIsReference(arg.type)) {
-//     const map: Record<string, AbiType> = {
-//       asset: AbiType.Asset,
-//       account: AbiType.Account,
-//       application: AbiType.Application,
-//     }
-//     return {
-//       name: arg.name,
-//       description: arg.description,
-//       type: map[arg.type],
-//     }
-//   }
-//
-//
-// }
-//
-//
-// const foo = (arg: algosdk.ABIArgumentType): AbiMethodArgumentDefinition => {
+
+const asMethodDefinitions = (appSpec: Arc32AppSpec): MethodDefinition[] => {
+  return appSpec.contract.methods.map((method) => {
+    const abiMethod = new algosdk.ABIMethod({
+      name: method.name,
+      desc: method.desc,
+      args: method.args,
+      returns: method.returns,
+    })
+    const signature = abiMethod.getSignature()
+    const hint = appSpec.hints ? appSpec.hints[signature] : undefined
+
+    return {
+      name: abiMethod.name,
+      signature: signature,
+      description: abiMethod.description,
+      arguments: abiMethod.args.map((arg, index) => ({
+        index: index + 1,
+        name: arg.name,
+        description: arg.description,
+        type: arg.type,
+        hint:
+          hint && arg.name && (hint.structs?.[arg.name] || hint.default_arguments?.[arg.name])
+            ? ({
+                struct: hint.structs?.[arg.name],
+                defaultValue: hint.default_arguments?.[arg.name],
+              } satisfies ArgumentHint)
+            : undefined,
+      })),
+      returns: {
+        ...abiMethod.returns,
+        hint:
+          hint && hint.structs?.['output']
+            ? {
+                struct: hint.structs?.['output'],
+              }
+            : undefined,
+      },
+    } satisfies MethodDefinition
+  })
+}
